deps/v8/test/inspector/debugger/get-possible-breakpoints-expected.txt

Test for Debugger.getPossibleBreakpoints

Running test: getPossibleBreakpointsInRange
Test start.scriptId != end.scriptId.
{
    error : {
        code : -32000
        message : Locations should contain the same scriptId
    }
    id : <messageId>
}
Test not existing scriptId.
{
    error : {
        code : -32000
        message : Script not found
    }
    id : <messageId>
}
Test end < start.
{
    id : <messageId>
    result : {
        locations : [
        ]
    }
}
Test empty range in first line.
{
    id : <messageId>
    result : {
        locations : [
        ]
    }
}
Test one character range in first line.
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 16
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
Test empty range in not first line.
{
    id : <messageId>
    result : {
        locations : [
        ]
    }
}
Test one character range in not first line.
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 16
                lineNumber : 1
                scriptId : <scriptId>
            }
        ]
    }
}
Test end is undefined
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 16
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 42
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 16
                lineNumber : 1
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 52
                lineNumber : 1
                scriptId : <scriptId>
            }
            [4] : {
                columnNumber : 54
                lineNumber : 1
                scriptId : <scriptId>
            }
            [5] : {
                columnNumber : 57
                lineNumber : 1
                scriptId : <scriptId>
            }
            [6] : {
                columnNumber : 0
                lineNumber : 2
                scriptId : <scriptId>
            }
        ]
    }
}
Test end.lineNumber > scripts.lineCount()
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 16
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 42
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 16
                lineNumber : 1
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 52
                lineNumber : 1
                scriptId : <scriptId>
            }
            [4] : {
                columnNumber : 54
                lineNumber : 1
                scriptId : <scriptId>
            }
            [5] : {
                columnNumber : 57
                lineNumber : 1
                scriptId : <scriptId>
            }
            [6] : {
                columnNumber : 0
                lineNumber : 2
                scriptId : <scriptId>
            }
        ]
    }
}
Test one string
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 16
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 42
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
Test end.columnNumber > end.line.length(), should be the same as previous.
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 16
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 42
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}

Running test: getPossibleBreakpointsInArrow
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 17
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 53
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 56
                lineNumber : 0
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 69
                lineNumber : 0
                scriptId : <scriptId>
            }
            [4] : {
                columnNumber : 71
                lineNumber : 0
                scriptId : <scriptId>
            }
            [5] : {
                columnNumber : 84
                lineNumber : 0
                scriptId : <scriptId>
            }
            [6] : {
                columnNumber : 90
                lineNumber : 0
                scriptId : <scriptId>
            }
            [7] : {
                columnNumber : 92
                lineNumber : 0
                scriptId : <scriptId>
            }
            [8] : {
                columnNumber : 94
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}

Running test: arrowFunctionFirstLine
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 18
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 47
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 49
                lineNumber : 0
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 51
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 18
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 47
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 49
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 51
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
paused in foo1
{
    columnNumber : 18
    lineNumber : 0
    scriptId : <scriptId>
}
paused in foo1
{
    columnNumber : 51
    lineNumber : 0
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 47
    lineNumber : 0
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 49
    lineNumber : 0
    scriptId : <scriptId>
}

Running test: arrowFunctionOnPause
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 0
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 28
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 57
                lineNumber : 0
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 59
                lineNumber : 0
                scriptId : <scriptId>
            }
            [4] : {
                columnNumber : 61
                lineNumber : 0
                scriptId : <scriptId>
            }
            [5] : {
                columnNumber : 18
                lineNumber : 1
                scriptId : <scriptId>
            }
            [6] : {
                columnNumber : 47
                lineNumber : 1
                scriptId : <scriptId>
            }
            [7] : {
                columnNumber : 49
                lineNumber : 1
                scriptId : <scriptId>
            }
            [8] : {
                columnNumber : 51
                lineNumber : 1
                scriptId : <scriptId>
            }
            [9] : {
                columnNumber : 0
                lineNumber : 2
                scriptId : <scriptId>
            }
            [10] : {
                columnNumber : 0
                lineNumber : 3
                scriptId : <scriptId>
            }
            [11] : {
                columnNumber : 6
                lineNumber : 3
                scriptId : <scriptId>
            }
        ]
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 0
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 28
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 57
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 59
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 61
            lineNumber : 0
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 18
            lineNumber : 1
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 47
            lineNumber : 1
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 49
            lineNumber : 1
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 51
            lineNumber : 1
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 0
            lineNumber : 2
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 0
            lineNumber : 3
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 6
            lineNumber : 3
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
paused in 
{
    columnNumber : 0
    lineNumber : 2
    scriptId : <scriptId>
}
paused in foo3
{
    columnNumber : 28
    lineNumber : 0
    scriptId : <scriptId>
}
paused in foo3
{
    columnNumber : 61
    lineNumber : 0
    scriptId : <scriptId>
}
paused in 
{
    columnNumber : 0
    lineNumber : 3
    scriptId : <scriptId>
}
paused in foo4
{
    columnNumber : 18
    lineNumber : 1
    scriptId : <scriptId>
}
paused in foo4
{
    columnNumber : 51
    lineNumber : 1
    scriptId : <scriptId>
}
paused in 
{
    columnNumber : 6
    lineNumber : 3
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 57
    lineNumber : 0
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 59
    lineNumber : 0
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 47
    lineNumber : 1
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 49
    lineNumber : 1
    scriptId : <scriptId>
}

Running test: getPossibleBreakpointsInRangeWithOffset
Test empty range in first line.
{
    id : <messageId>
    result : {
        locations : [
        ]
    }
}
Test one character range in first line.
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 17
                lineNumber : 1
                scriptId : <scriptId>
            }
        ]
    }
}
Test empty range in not first line.
{
    id : <messageId>
    result : {
        locations : [
        ]
    }
}
Test one character range in not first line.
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 16
                lineNumber : 2
                scriptId : <scriptId>
            }
        ]
    }
}
Test end is undefined
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 17
                lineNumber : 1
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 43
                lineNumber : 1
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 16
                lineNumber : 2
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 52
                lineNumber : 2
                scriptId : <scriptId>
            }
            [4] : {
                columnNumber : 54
                lineNumber : 2
                scriptId : <scriptId>
            }
            [5] : {
                columnNumber : 57
                lineNumber : 2
                scriptId : <scriptId>
            }
            [6] : {
                columnNumber : 0
                lineNumber : 3
                scriptId : <scriptId>
            }
        ]
    }
}
Test end.lineNumber > scripts.lineCount()
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 17
                lineNumber : 1
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 43
                lineNumber : 1
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 16
                lineNumber : 2
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 52
                lineNumber : 2
                scriptId : <scriptId>
            }
            [4] : {
                columnNumber : 54
                lineNumber : 2
                scriptId : <scriptId>
            }
            [5] : {
                columnNumber : 57
                lineNumber : 2
                scriptId : <scriptId>
            }
            [6] : {
                columnNumber : 0
                lineNumber : 3
                scriptId : <scriptId>
            }
        ]
    }
}
Test one string
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 17
                lineNumber : 1
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 43
                lineNumber : 1
                scriptId : <scriptId>
            }
        ]
    }
}
Test end.columnNumber > end.line.length(), should be the same as previous.
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 17
                lineNumber : 1
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 43
                lineNumber : 1
                scriptId : <scriptId>
            }
        ]
    }
}

Running test: withOffset
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 36
                lineNumber : 3
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 65
                lineNumber : 3
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 67
                lineNumber : 3
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 69
                lineNumber : 3
                scriptId : <scriptId>
            }
            [4] : {
                columnNumber : 18
                lineNumber : 4
                scriptId : <scriptId>
            }
            [5] : {
                columnNumber : 47
                lineNumber : 4
                scriptId : <scriptId>
            }
            [6] : {
                columnNumber : 49
                lineNumber : 4
                scriptId : <scriptId>
            }
            [7] : {
                columnNumber : 51
                lineNumber : 4
                scriptId : <scriptId>
            }
        ]
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 36
            lineNumber : 3
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 65
            lineNumber : 3
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 67
            lineNumber : 3
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 69
            lineNumber : 3
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 18
            lineNumber : 4
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 47
            lineNumber : 4
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 49
            lineNumber : 4
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
{
    id : <messageId>
    result : {
        actualLocation : {
            columnNumber : 51
            lineNumber : 4
            scriptId : <scriptId>
        }
        breakpointId : <breakpointId>
    }
}
paused in foo5
{
    columnNumber : 36
    lineNumber : 3
    scriptId : <scriptId>
}
paused in foo5
{
    columnNumber : 69
    lineNumber : 3
    scriptId : <scriptId>
}
paused in foo6
{
    columnNumber : 18
    lineNumber : 4
    scriptId : <scriptId>
}
paused in foo6
{
    columnNumber : 51
    lineNumber : 4
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 65
    lineNumber : 3
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 67
    lineNumber : 3
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 47
    lineNumber : 4
    scriptId : <scriptId>
}
paused in Promise.resolve.then
{
    columnNumber : 49
    lineNumber : 4
    scriptId : <scriptId>
}

Running test: arrowFunctionReturn
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 0
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 6
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 9
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 34
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 45
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 48
                lineNumber : 0
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 49
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 0
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 8
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 12
                lineNumber : 0
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 13
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 17
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 21
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 22
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 0
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 6
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 8
                lineNumber : 0
                scriptId : <scriptId>
            }
            [3] : {
                columnNumber : 9
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
{
    id : <messageId>
    result : {
        locations : [
            [0] : {
                columnNumber : 0
                lineNumber : 0
                scriptId : <scriptId>
            }
            [1] : {
                columnNumber : 8
                lineNumber : 0
                scriptId : <scriptId>
            }
            [2] : {
                columnNumber : 19
                lineNumber : 0
                scriptId : <scriptId>
            }
        ]
    }
}
